medical-qc-backend/python_model/requirements.txt
nibabel
numpy
opencv-python
Pillow
pydicom
SimpleITK
torch
torchvision
websockets
